Fix batch of C# compilation errors in scanner modules
- proc.Name → proc.ProcessName in AltV, FiveM, RageMP, GameReplay modules
- proc.MainModule → proc.MainModule?.FileName for type safety
- CheatLoaderInjectorForensicScanModule: fix mixed void/Task returns in lambda
- Dota2CheatScanModule: move yield return out of try/catch (CS1626)
- EtwTamperScanModule: fix QueryTrace(0UL) for ulong parameter
- GameMemoryReadAccessScanModule: DuplicateHandle return nint→bool (CS0029)
- InlineHookDetectionScanModule: fix uint+int arithmetic (CS0266)
- HardwareAimbotDeviceForensicScanModule: fix StreamReader named params (CS1739)
- AntiCheatStatusScanModule: Array.Empty<string>() for implicit array (CS0826)
- RemoteAccessToolScanModule: Array.Empty<string>() for implicit array (CS0826)
- Add System.ServiceProcess.ServiceController NuGet package (CS1069)
- ApcInjectionScanModule: replace goto-into-finally with if block (CS0159)
- SuspiciousImportedFunctionsScanModule: add ct param to ScanProcess (CS0103)
- GamingVPNBanEvasionForensicScanModule: fix Path.Combine named param (CS1739)
- GameDirectoryInjectionScanModule: fix HashUtil namespace + method (CS0234)
- BootSectorScanModule: add missing IntPtr.Zero to ReadFile call (CS7036)

+++ scanner/src/ZeroTrace.Core/Modules/InlineHookDetectionScanModule.cs
@@ -253,8 +253,8 @@ public sealed class InlineHookDetectionScanModule : IScanModule
             {
                 try
                 {
-                    uint nameRvaOff = addrOfNames + i * 4;
-                    if (nameRvaOff + 4 > expRead) continue;
+                    uint nameRvaOff = addrOfNames + i * 4U;
+                    if (nameRvaOff + 4U > (uint)expRead) continue;
                     uint nameRva = BitConverter.ToUInt32(expBuf, (int)nameRvaOff) - exportRva;
                     if (nameRva >= expRead) continue;
 
@@ -268,11 +268,11 @@ public sealed class InlineHookDetectionScanModule : IScanModule
                     if (!HighValueExports.Contains(funcName)) continue;
 
                     // Get ordinal → function RVA
-                    uint ordOff = addrOfOrds + i * 2;
-                    if (ordOff + 2 > expRead) continue;
+                    uint ordOff = addrOfOrds + i * 2U;
+                    if (ordOff + 2U > (uint)expRead) continue;
                     ushort ord = BitConverter.ToUInt16(expBuf, (int)ordOff);
-                    uint funcRvaOff = addrOfFuncs + ord * 4;
-                    if (funcRvaOff + 4 > expRead) continue;
+                    uint funcRvaOff = addrOfFuncs + (uint)ord * 4U;
+                    if (funcRvaOff + 4U > (uint)expRead) continue;
                     uint funcRva = BitConverter.ToUInt32(expBuf, (int)funcRvaOff);
                     if (funcRva == 0) continue;
